### 2023-03-24 f6c7636
closes #13; renamed shape_masks and grating_masks in checkerboard
