### 2023-03-24 02d0081
closes #42; added clipping and fixed some defaults
