### 2023-03-24 51cbe75
closes #51; removed stripe intensities from yazdanbakhsh
