### 2023-03-27 de9c765
removed non-existent angular-components-functions from demo; added segments; fixed bugs in segments
