+++ docs/reference/demos/components/angulars.md
@@ -38,11 +38,12 @@ from stimupy.components.angulars import wedge
 w_height = iw.IntSlider(value=10, min=1, max=20, description="height [deg]")
 w_width = iw.IntSlider(value=10, min=1, max=20, description="width [deg]")
 w_ppd = iw.IntSlider(value=20, min=1, max=40, description="ppd")
+w_rot = iw.IntSlider(value=0, min=0, max=360, description="rotation [deg]")
 
-w_wwidth = iw.IntSlider(value=45, min=1, max=90, description="width [deg]")
+w_wwidth = iw.IntSlider(value=45, min=1, max=90, description="angle [deg]")
+
 w_oradius = iw.FloatSlider(value=3, min=1, max=6, description="outer radius [deg]")
 w_iradius = iw.FloatSlider(value=0, min=0, max=3, description="inner radius [deg]")
-w_rot = iw.IntSlider(value=0, min=0, max=360, description="rotation [deg]")
 
 w_int = iw.FloatSlider(value=0.5, min=0, max=1, description="intensity wedge")
 w_int_back = iw.FloatSlider(value=0., min=0, max=1, description="intensity background")
@@ -51,11 +52,12 @@ w_ori = iw.Dropdown(value="mean", options=['mean', 'corner', 'center'], descript
 w_mask = iw.ToggleButton(value=False, disabled=False, description="add mask")
 
 # Layout
-b_im_size = iw.HBox([w_height, w_width, w_ppd])
-b_geometry = iw.HBox([w_wwidth, w_oradius, w_iradius, w_rot])
+b_im_size = iw.HBox([w_height, w_width, w_ppd, w_rot])
+b_angles = iw.HBox([w_wwidth])
+b_geometry = iw.HBox([w_oradius, w_iradius])
 b_intensities = iw.HBox([w_int, w_int_back])
 b_add = iw.HBox([w_ori, w_mask])
-ui = iw.VBox([b_im_size, b_geometry, b_intensities, b_add])
+ui = iw.VBox([b_im_size, b_angles, b_geometry, b_intensities, b_add])
 
 # Function for showing stim
 def show_wedge(
@@ -74,7 +76,7 @@ def show_wedge(
     stim = wedge(
         visual_size=(height, width),
         ppd=ppd,
-        width=wwidth,
+        angle=wwidth,
         radius=radius,
         rotation=rotation,
         inner_radius=inner_radius,
@@ -106,146 +108,81 @@ out = iw.interactive_output(
 display(ui, out)
 ```
 
-## Grating
-{py:func}`stimupy.components.angulars.grating`
+## Segments
+{py:func}`stimupy.components.angulars.segments`
 
 ```{code-cell} ipython3
-from stimupy.components.angulars import grating
+from stimupy.components.angulars import segments
 
 # Define widgets
 w_height = iw.IntSlider(value=10, min=1, max=20, description="height [deg]")
 w_width = iw.IntSlider(value=10, min=1, max=20, description="width [deg]")
 w_ppd = iw.IntSlider(value=20, min=1, max=40, description="ppd")
-
-w_nseg = iw.IntSlider(value=6, min=2, max=12, description="n_segments")
 w_rot = iw.IntSlider(value=0, min=0, max=360, description="rotation [deg]")
 
-w_int1 = iw.FloatSlider(value=1, min=0, max=1, description="intensity1")
-w_int2 = iw.FloatSlider(value=0., min=0, max=1, description="intensity2")
+w_wwidth1 = iw.IntSlider(value=45, min=1, max=90, description="angle1 [deg]")
+w_wwidth2 = iw.IntSlider(value=90, min=1, max=180, description="angle2 [deg]")
[…]
