+++ stimupy/components/radials.py
@@ -127,8 +127,15 @@ def rings(
         visual_size = [x for x in visual_size if x is not None]
         visual_size = resolution.validate_visual_size(visual_size)
 
-    if np.diff(radii).min() < 0:
-        raise ValueError("radii need to monotonically increase")
+    if not isinstance(radii, (int, float)):
+        if np.diff(radii).min() < 0:
+            raise ValueError("radii need to monotonically increase")
+    else:
+        radii = (radii,)
+    if isinstance(intensity_rings, (int, float)):
+        ints = (intensity_rings,)
+    else:
+        ints = intensity_rings
 
     # Get masks for rings
     params = mask_rings(radii=radii, shape=shape, visual_size=visual_size, ppd=ppd, origin=origin)
@@ -139,7 +146,7 @@ def rings(
     distances = base["radial"]
 
     img = np.ones(shape) * intensity_background
-    ints = [*itertools.islice(itertools.cycle(intensity_rings), len(radii))]
+    ints = [*itertools.islice(itertools.cycle(ints), len(radii))]
     for radius, intensity in zip(reversed(radii), reversed(ints)):
         img[distances < radius] = intensity
 
