### 2023-03-27 af053df
increased frequency in waves-demo because of angular-wave
