+++ stimupy/components/frames.py
@@ -92,7 +92,7 @@ def frames(
     """
     if radii is None:
         raise ValueError("frames() missing argument 'radii' which is not 'None'")
-    
+
     if not isinstance(radii, (int, float)):
         if np.diff(radii).min() < 0:
             raise ValueError("radii need to monotonically increase")
