### 2023-03-27 db6a3d4
fixed args in mondrians demos, and fixed bug in mondrians
