### 2023-03-27 941cbaa
updated some sbcs-demos args; and fixed bugs in dotted sbcs
