### 2023-03-28 530dbc0
docs(landing page): use description from README, more detailed cards
