+++ docs/index.md
@@ -1,28 +1,62 @@
 # stimupy
 
-Welcome to `stimupy`
+```{include} ../README.md
+:start-line: 3
+:end-line: 7
+```
+
+```{include} ../README.md
+:start-line: 17
+:end-line: 30
+```
 
 ::::{grid} 2
 :gutter: 3
 
-:::{grid-item-card} Get started 🏃
+:::{grid-item-card} Get started 
 :link: getting_started/getting_started
 :link-type: doc
+:text-align: center
+🏃
+^^^
+
+Install `stimupy`
+and follow an introductory tutorial
 :::
 
-:::{grid-item-card} Learn more 🧑‍🏫
+:::{grid-item-card} Learn more
 :link: topic_guides/topic_guides
 :link-type: doc
+:text-align: center
+🧑‍🏫
+^^^
+
+Learn more about design decisions
+and particular conventions and topics within `stimupy`
 :::
 
-:::{grid-item-card} Reference 📑
+:::{grid-item-card} Reference
 :link: reference/api
 :link-type: doc
+:text-align: center
+📑
+^^^
+
+Look up functions and their specific parameters,
+including through interactive demos
+
 :::
 
-:::{grid-item-card} Get in touch 📨, contribute 🎁 
+:::{grid-item-card} Get in touch, or contribute
 :link: contributing/get_in_touch
 :link-type: doc
-:::
+:text-align: center
+📨, 🎁
+^^^
 
+Ask a question,
+report a bug,
+request a feature,
+or contribute.
+:::
 ::::
