### 2023-03-28 506ebf6
docs: Detail on dev installation instructions
Closes #19
