+++ .github/ISSUE_TEMPLATE/stimulus-request.md
@@ -1,0 +1,33 @@
+---
+name: Stimulus request
+about: Suggest a stimulus function
+title: Add <STIMULUS>
+labels: enhancement
+assignees: ''
+
+---
+
+## What stimulus would you like?
+A clear and concise description of what it should look like.
+Please attach images/drawings to illustrate.
+If a citation for this stimulus is available, please also mention it.
+
+## What are the relevant stimulus parameters?
+Since `stimupy` stimuli are fully parameterizable,
+please briefly describe what parameters you'd like this new function to have,
+and what their expected behavior is.
+
+*NOTE*: if you want only a single parameterization (aside from resolution),
+that does not fit within the scope a new stimulus function
+-- instead, please request a new stimulus *set*.
+
+## Is the stimulus is similar to an existing `stimupy` stimulus?
+If so, please describe the *differences* between the existing stimulus,
+and the proposed one
+-- especially the difference in behavior for parameters.
+
+## Describe alternatives you've considered
+A clear and concise description of any alternative solutions or features you've considered.
+
+## Additional context
+Add any other context or screenshots about the feature request here.
