+++ .github/ISSUE_TEMPLATE/stimulus-set-request.md
@@ -1,0 +1,17 @@
+---
+name: Stimulus set request
+about: Suggest a stimulus set
+title: Add set <CITATION>
+labels: enhancement
+assignees: ''
+
+---
+
+## What stimulus *set* would you like?
+Please provide a DOI / full bibliographical reference to the publication that describes this stimulus set.
+
+## Are the additional parameters/aspects of the stimuli, not described in the publication?
+If so, please share
+
+## Do you have or know of any implementation of this stimulus set?
+If so, please share if possible.
