+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,28 @@
+---
+name: Bug report
+about: Report an error or unexpected behavior
+title: ''
+labels: bug
+assignees: ''
+
+---
+
+## Context
+Describe what you were attempting to do
+
+## Minimal Reproducible Example
+Provide a (minimal) piece of code that leads to the bug.
+Ideally, this piece of code produces the bug every time,
+especially in a "fresh" Python session after restart
+
+## Expected behavior
+A clear and concise description of what you expected to happen.
+
+## Observed behavior
+A clear and concise description of the error or unexpected behavior that happens instead.
+Attach screenshots if helpful to illustrate.
+
+## System information:
+ - OS and version: [e.g. macOS 12, Ubuntu 22.04]
+ - Python version: [e.g. 3.11]
+ - stimupy version: [e.g. 1.0.0]
