+++ .github/ISSUE_TEMPLATE/documentation-change.md
@@ -1,0 +1,25 @@
+---
+name: Documentation change
+about: Suggest a correction or change to documentation
+title: ''
+labels: documentation
+assignees: ''
+
+---
+
+## Is the proposed change to a specific (function) *docstring*?
+If so, please specify which function: [e.g. stimupy.components.shapes.rectangle]
+
+## Is the proposed change to a specific function *demo*?
+If so, please specify which function: [e.g. stimupy.components.shapes.rectangle]
+
+## Is the proposed change to a specific page?
+If so, please specify which page: [e.g., https://stimupy.readthedocs.io/en/latest/index.html]
+
+## Describe the change
+If specific lines/words should be changed, please provide current & new lines, e.g.:
+- "this is the old line"
++ "this is the new line"
+
+If the proposed change/edit/clarification is more general,
+please describe.
