+++ docs/contributing/stimulus.md
@@ -1,1 +1,0 @@
-# Contribute a parameterizable stimulus function
