### 2023-03-29 31c4ae6
docs(tutorial): More verbose tutorial component
