+++ docs/getting_started/first_stim.md
@@ -20,89 +20,62 @@ import numpy as np
 
 Here you will be introduced to the basics of `stimupy`,
 and create your first stimulus through this package.
-At its core, `stimupy` provides a large number of functions
+At its core, `stimupy` provides a large number of *functions*
 that each draw some kind of (visual) component in a `numpy.ndarray`.
 
 ```{tip} Launch on Binder
 This page can also be launched as an interactive Jupyter Notebook on Binder -- see icon at top
 ```
 
-## Basic shapes
+## Drawing a basic shape
 
 The most basic stimuli that `stimupy` provides, are basic geometric shapes.
-These can be found in `stimupy.components.shapes`.
-
+These functions be found in the `stimupy.components.shapes` *module*.
+To be able to access these,
+we first have to *import* this module
+into our current Python session:
 ```{code-cell}
 from stimupy.components import shapes
 ```
+
+`````{margin}
+````{tip}
+In some interactive Python interpreters,
+such as IPython, or Jupyter Notebooks,
+you can also see all contents of a module
+by typing the (imported) module name
+and pressing <TAB> to autocomplete:
+```
+>>> shapes.<TAB>
+```
+````
+`````
+This module contains the following functions:
 ```{code-cell}
 :tags: [hide-input]
 print(f"Available basic shapes: {shapes.__all__}")
 ```
 
-Each of these is a separate function
-that takes in parameters specifying the geometry of the shape.
+Each of these is a separate function.
 Let's take a look at one of these:
-
 ```{code-cell}
-:tags: [hide-input]
-help(shapes.rectangle)
+stim = shapes.rectangle(visual_size=(6,8), ppd=10, rectangle_size=(4,4))
 ```
 
 This function creates an image, and draws a rectangle inside it.
 Everywhere the rectangle isn't, is considered the "background".
+It returns a `dict`ionary,
+mapping `str`ings as *keys*, to all kinds of *values*.
 
-The arguments consist of three main categories:
-
-- image size & resolution:
-  - `visual_size` (height, width) of the whole image, in degrees visual angle
-  - pixels per degree (`ppd`) of visual angle
-  - `shape` in pixels
-  - because these are three are inter-dependent, only two out of 3 need to be specified
-    (see [resolution guide](../topic_guides/resolution) for more information)
-- component geometry:
-  - `rectangle_size` in degrees visual angle
-  - `rectangle_position`, as distance from top-left corner, in degrees visual angle
-- intensity ("photometric"):
-  - in this case: `intensity_rectangle` and `intensity_background`
-  - all `stimupy` stimuli by default are in the range [0, 1]
-
-```{code-cell}
-stim = shapes.rectangle(visual_size=(6,8), ppd=10, rectangle_size=(4,4))
-```
-```{code-cell}
-:tags: [remove-cell]
-glue('type', type(stim))
-glue('keys',stim.keys())
-glue('imgtype',type(stim["img"]))
-```
-
-This is output is a {glue:}`type`, with {glue:}`keys`.
-
-The generated stimulus image is also in this output {glue:}`type`, under the key `"img"`.
-As you can see, this output also contains all of the parameters for the component,
-including those for which default values were used.
-Thus, this output is self-documenting:
-all the necessary information to produce this stimulus are in the stimulus.
-
-The `"img"` is a {glue:}`imgtype`,
+One of these values in in this output `dict`,
+is the generated stimulus image is under the key `"img"`.
+This `"img"` is a `numpy.ndarray`,
 where each entry in this array corresponds to a pixel in the image.
-Thus, the `shape` of the array is the (specified) `shape` in pixels of the image.
-
-```{code-cell}
-:tags: [hide-input]
[…]
