+++ docs/getting_started/getting_started.md
@@ -1,14 +1,15 @@
-# Getting started with `stimupy`
+# Getting started (Tutorial)
 
 The following pages serve as **tutorial**,
-walking you through the very basics of installing and using `stimupy`.
+walking you through the very basics of using `stimupy`.
 
 Along the way, they also refer to some [topic guides](../topic_guides/),
 which have more in-depth explanation on various concepts and design decisions.
 
+First, [install `stimupy`](./installation/)
+
 ```{toctree}
 :numbered:
-installation
 first_stim
 composition
 stimulus
