### 2023-03-29 506274e
docs(tutorial): Fix a reference
