+++ docs/getting_started/getting_started.md
@@ -3,7 +3,7 @@
 The following pages serve as **tutorial**,
 walking you through the very basics of using `stimupy`.
 
-Along the way, they also refer to some [topic guides](../topic_guides/),
+Along the way, they also refer to some [topic guides](../topic_guides/topic_guides),
 which have more in-depth explanation on various concepts and design decisions.
 
 First, [install `stimupy`](./installation/)
