### 2023-03-29 #5301 @billbrod
@emdupre I am reviewing another paper for JOSS at the moment, and don't think I can do a second. Sorry, this package looks really interesting!
