### 2023-03-30 a1d5e18
build: Use PSR to manage version number
