### 2023-03-30 7123594
build: GHA workflow for bumping version
