+++ .github/workflows/version.yml
@@ -1,0 +1,29 @@
+name: Bump version (if necessary)
+
+on: push
+
+jobs:
+  bump-version:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout repository 
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: 0 # checkout full commit history
+
+      - name: Setup Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: 3.11
+
+      - name: Configure git to be able to push to repo
+        env:
+          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        run: |
+          git config user.name github-actions
+          git config user.email github-actions@github.com
+
+      - name: Run Semantic Release
+        run: |
+          pipx run python-semantic-release version -v DEBUG
