### 2023-03-30 3fc3606
build: GHA workflow for bumping version
