### 2023-03-30 b25eb57
build: Create bump-commit as well
Otherwise the updated version numbers in pyproject.toml and __init__.py won't actually be commited
