+++ pyproject.toml
@@ -74,3 +74,5 @@ version_variable = [                    # version location(s)
   "stimupy/__init__.py:__version__"     # in the init.py, so that stimupy.__version__ works
 ]
 patch_without_tag = false               # if nothing triggers a patch, bump?
+commit_version_number = true            # commit changes to version number?
+commit_subject = "v{version}"
