### 2023-03-30 510e9e1
build: only bump on PR merge
