### 2023-03-30 bf03dc2
build: only bump on PR merge
