+++ .github/workflows/version.yml
@@ -1,9 +1,13 @@
 name: Bump version (if necessary)
 
-on: push
+on:
+  pull_request:
+    types:
+      - closed
 
 jobs:
   bump-version:
+    if: github.event.pull_request.merged == true
     runs-on: ubuntu-latest
 
     steps:
