### 2023-03-30 a660a4c
build: only bump on PR merge
and on manual trigger
