+++ .github/workflows/version.yml
@@ -1,9 +1,14 @@
 name: Bump version (if necessary)
 
-on: push
+on:
+  pull_request_target:
+    types:
+      - closed
+  workflow_dispatch:
 
 jobs:
   bump-version:
+    if: github.event.pull_request.merged == true
     runs-on: ubuntu-latest
 
     steps:
