### 2023-03-31 c2ab046
build: derive version from commit (not tag)
Because commits will be (more) unambiguous, and if multiple branches have different version-commits, that will need to be resolved in merge-conflicts anyway, thus providing a good check. The alternative, derive from tags, is more fragile: the same commit could have multiple version tags...
