+++ .github/workflows/version.yml
@@ -29,9 +29,10 @@ jobs:
           git config user.name github-actions
           git config user.email github-actions@github.com
 
-      - name: Run Semantic Release
+      - name: Bump version using Semantic Release
         env:
           GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         run: |
           pipx run python-semantic-release version -v DEBUG
-          git push
+          VERSION="v"$(pipx run python-semantic-release print-version --current)
+          git push --atomic origin HEAD $VERSION
