### 2023-03-31 9be4db0
docs(tutorial): more detail on composition, masks
Closes #70
