+++ docs/getting_started/composition.md
@@ -12,57 +12,134 @@ kernelspec:
 
 # Composing stimuli, composed stimuli
 
-## Masked regions
+Most stimuli consist not just of one shape or element,
+but as a composition of multiple components.
+The geometric {py:mod}`components <stimupy.components>`
+form the basic building blocks for all stimuli implemented in `stimupy`.
+In this tutorial, we will explore how stimulus consisting of multiple geometric elements,
+can be composed using the functions that generate components.
 
-Another key in the stimulus-`dict` not addressed yet, is the `"mask"`.
-This too is a `numpy.ndarray`, with the same shape as `"img"`
-(i.e., each entry corresponds to a pixel in `"img"`).
-
-However, rather than the values here being any floating point pixel-intensities
-(default in range $[0, 1]$),
-the `"mask"` contains integer-values.
-Each integer-value corresponds to a geometric region of interest.
-For basic shapes like these there are only two such regions:
-the background (mask value: `0`), and the shape itself (mask value: `1`).
-These can be used to *mask* the regions: all pixels with value `1` belong to the shape.
-
-We can visualize these as well,
-overlayed as colorcoding on top of the stimulus:
 ```{code-cell}
-plot_stim(stim, mask='shape_mask')
+import numpy as np
+import matplotlib.pyplot as plt
+
+from stimupy.utils import plot_stim, plot_stimuli
+from stimupy.components import shapes
+```
+
+First we will create our two components that we want to combine in some way;
+here we will use the {py:func}`rectangle <stimupy.components.shapes.rectangle>`
+and {py:func}`disc <stimupy.components.shapes.disc>`
+(from our [previous tutorial](first_stim))
+
+```{code-cell}
+rectangle = shapes.rectangle(visual_size=(6,8), ppd=10,
+                        rectangle_size=(4,2), rectangle_position=(1,2),
+                        intensity_rectangle=.7)
+
+disc = shapes.disc(visual_size=(6,8), ppd=10,
+                        radius=2,
+                        intensity_disc=1, intensity_background=.5)
+
+plot_stimuli({"rectangle": rectangle, "disc": disc})
 plt.show()
 ```
 
-## Composition
+Since the `"img"`s of the two stimuli are {py:class}`numpy.ndarray`s,
+we can simple manipulate these to combine our stimuli:
+```{code-cell}
+new_img = disc["img"] - rectangle["img"]
 
-Another use of region masking
-is to compose more complicated stimuli
-from basic components.
+plt.imshow(new_img, cmap="gray")
+```
+
+We can even turn this into a something resembling more of a `stimupy` stimulus
+-- and use the `stimupy` tooling
+like {py:func}`plot_stimuli <stimupy.utils.plot_stimuli>` on it --
+simply by wrapping it in a {py:class}`dict`,
+and optionally adding some additional metadata:
+```{code-cell}
+new_stim = {
+  "img": new_img,
+  "visual_size": disc["visual_size"],
+  "ppd": disc["ppd"]
+}
+
+plot_stimuli({"rect": rectangle, "disc": disc, "composed": new_stim})
+```
+## Masked regions
+
+The downside of this, is that such an operation (`-`) on the `"img"`s
+uses the *whole* `imgs`s, including the background.
+In other words, these operations aren't *content aware*;
+they treat all pixels the same, rather than restricting operations
+to the regions that we care about (e.g., the geometric shapes).
+We'd like to be to have a bit more control,
+for instance subtracting just a where the shapes overlap.
+
+One key in the stimulus-{py:class}`dict`s not addressed yet, is the `mask`.
[…]
