### 2023-03-31 8889e56
build(vesion): don't auto-bump to `v1.0.0`
All BREAKING CHANGES will bump minor instead of major version
