### 2023-03-31 #5301 @emdupre
👋 Hi @JonathanReardon and @mdcutone,

Would you be willing to review _stimupy: A Python stimulus creation package for vision science_ for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org/)?

You can learn more about reviewing for JOSS -- including our conflict of interest policy -- [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The review takes place on GitHub and focuses on [the software](https://github.com/computational-psychology/stimupy) and this short paper: https://github.com/openjournals/joss-papers/blob/joss.05301/joss.05301/10.21105.joss.05301.pdf

If you're available to review this work, please let me know and I'll add you as a reviewer. Once I have sufficient reviewers I'll open a dedicated review issue.

Thank you for considering !
