### 2023-03-31 #5321 @editorialbot
[REVIEW]: stimupy: A Python stimulus creation package for vision science
**Submitting author:** @lynnschmittwilken (<a href="http://orcid.org/0000-0003-3621-9576">Lynn Schmittwilken</a>)
**Repository:** https://github.com/computational-psychology/stimupy
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
**Archive:** 10.17605/OSF.IO/Z439V
