### 2023-03-31 #5321 @emdupre
Hi @alexander-pastukhov and @JonathanReardon 👋 Thank you again for agreeing to review this submission ! The review will take place in this issue, and you can generate your individual reviewer checklists by asking editorialbot directly with `@editorialbot generate my checklist`.

In working through the checklist, you're likely to have specific feedback on _stimupy._ Whenever possible, please open relevant issues on the [linked software repository](https://github.com/computational-psychology/stimupy) (and cross-link them with this issue) rather than discussing them here. This helps to make sure that feedback is translated into actionable items to improve the software !

If you aren't sure how to get started, please see the [Reviewing for JOSS guide](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) -- and, of course, feel free to ping me with any questions !
