+++ stimupy/components/__init__.py
@@ -184,6 +184,24 @@ def mask_regions(
     }
 
 
+def combine_masks(*masks):
+    # Initialize
+    combined_mask = np.zeros_like(masks[0])
+    for mask in masks:
+        # Check that masks have the same shape
+        if not mask.shape == combined_mask.shape:
+            raise ValueError("Not all masks have the same shape")
+
+        # Check that masks don't overlap
+        if (combined_mask & mask).any():
+            raise ValueError("Masks overlap")
+
+        # Combine: increase `mask`-idc by adding the current highest idx in combined_mask
+        combined_mask = np.where(mask, mask + combined_mask.max(), combined_mask)
+
+    return combined_mask
+
+
 def draw_regions(mask, intensities, intensity_background=0.5):
     """Draw regions defined by mask, with given intensities
 
