### 2023-04-03 725eaf0
docs: docstring for `mask_regions`
