### 2023-04-03 2d2cad2
docs: docstring for `draw_regions`
