### 2023-04-03 c1aa531
docs: docstring for `combine_masks`
