### 2023-04-04 8e17857
refactor(image_base): `rotation` consistent, extra `oblique_y`
`rotation` argument now consistently rotations counterclockwise from 3 o'clock position (in line with mathematical convention, i.e., unit circle)

Also added an `oblique_y` metric, orthogonal to `oblique` (which is implicit `oblique_x`)
