### 2023-04-04 a7f8d56
refactor: update callers of `oblique` / `rotation`
