+++ stimupy/stimuli/checkerboards.py
@@ -220,7 +220,7 @@ def checkerboard(
         n_phases=board_shape[1],
         phase_width=check_visual_size[1],
         period=period,
-        rotation=rotation + 90,
+        rotation=rotation - 90,
         phase_shift=0,
         intensities=intensity_checks,
         origin="corner",
@@ -257,7 +257,7 @@ def checkerboard(
             n_phases=board_shape[1],
             phase_width=check_visual_size[1],
             period=period,
-            rotation=rotation + 90,
+            rotation=rotation - 90,
             phase_shift=0,
             intensities=intensity_checks,
             origin="corner",
