### 2023-04-04 b441a34
docs: consistent docstrings & defaults for `rotation` arguments
