### 2023-04-04 computational-psychology/stimupy#69 @alexander-pastukhov
Issue on page /getting_started/stimulus.html
3.1 Code #5 `plot_stim(two_sided_stim, mask='target_mask') plt.show()` fails with "TypeError: __init__() got an unexpected keyword argument 'extend'"
3.2 Todorovic illusion output shows error "AttributeError: module 'stimupy.stimuli.todorovics' has no attribute 'two_sided_rectangle'" in the docs themselves (also when I run it).
3.3. White’s effect stimulus  - no output in docs but running it in Jupyter notebook generates "TypeError: white_two_rows() got an unexpected keyword argument 'target_height'".
