### 2023-04-04 computational-psychology/stimupy#70 @alexander-pastukhov
Issue on page /getting_started/first_stim.html
2.2. Masked regions: KeyError: 'shape_mask'. Same applies to first code chunk in 2.3 Composition. However, the second code in 2.3 that is supposed to show different masks fails with KeyError: 'visual_size'. If this code is not intended to run, this should be either explicitly stated or, better still, this should be a working example.
