### 2023-04-04 computational-psychology/stimupy#71 @alexander-pastukhov
Issue on page /getting_started/installation.html
PyPI installation instructions are confusing, I read the "is not (yet) available" bit only _after_ I pip install failed. Hide it until it is on PyPI?
