### 2023-04-04 computational-psychology/stimupy#72 @JonathanReardon
problem installing
Hello, I'm reviewing stimupy for JOSS.

I'm having trouble installing the package. I've tried the suggested pip install (below) along with 'pip3 install..', 'python3 -m pip install..'. All install with no errors but I am unable to import the package.

```
pip install https://github.com/computational-psychology/stimupy/zipball/main
```

Do you have any suggestions? I'm using Python 3.10.6, Linux (Pop) OS.
