### 2023-04-04 computational-psychology/stimupy#73 @JonathanReardon
removal of unnecessary information
I would suggest removing the following part from the [documentation](https://stimupy.readthedocs.io/en/latest/getting_started/installation.html). I'm not sure what purpose this serves, and at first glance it looks like this method is available.

```
stimupy is not (yet) available on PyPI – but if it was, one could:
pip install stimupy
```

I'm also not sure why the 'Surely this must also be possible…' is on the Conda (install) page. Is the purpose to elicit help in setting that up?
