### 2023-04-04 4d62316
refactor(rectilinear): construct rectilinear from obliques
Both for consistency, and to make rectilinear distances rotatable.
