### 2023-04-04 0670d6b
feat(rectilinear): add rotation to rectilinear stimuli
