### 2023-04-04 computational-psychology/stimupy#70 @alexander-pastukhov
Same applies to first code chunk in 2.3 Composition. However, the second code in 2.3 that is supposed to show different masks fails with KeyError: 'visual_size'
