### 2023-04-04 computational-psychology/stimupy#74 @alexander-pastukhov
Issue on page /reference/demos/components/angulars.html
Please include imports that are present in the binder notebook, so that examples can be easily run in local Jupyter notebook without the need to figure them out. Also, it is fairly easy to find a combination of parameters that results in an intimidating looking error. I like informative error messages but for interactive use a less intimidating error output (i.e., just the last bit without the entire stack) would be advantages, I think.
