### 2023-04-04 #5321 @alexander-pastukhov
@emdupre I am first-time reviewer for JOSS, so I am not sure how do I proceed. I like the concept behind the package and I can see how it would be widely used by the visual community (I probably would with a caveat that virtually all my stimuli are dynamic and the package appears to only be able to handle the static ones). Paper-wise, I have only couple of minor issues (but I am not sure where and how I post them).

But... Either the docs, or the package (or both) are half-backed. There is a limited number of examples in "Getting started" and a lot of them simply do not work (and errors are fairly variable, implying that this is not the same issue such as a single missing dependence). An important section of documentation
[Topic Guides](https://stimupy.readthedocs.io/en/latest/topic_guides/topic_guides.html) is completely empty. Couple of unit test files produce quite a few warnings (17 for test_overviews, 6 for test_waves), which I don't think was the plan. API docs seem to be complete but without topic guides it is not very usable for a novice. "Demos of parameters" section also seem to work but I would perhaps polish the error handling, so it looks less as a complete crash (which is good for experimental code but not so ideal in an interactive tool).

In short, is it a good package that will be useful for the community - yes, should it be accepted by JOSS once it is more polished and ready - absolutely, is it ready (at least the docs) - no.
