### 2023-04-05 2fb5f42
refactor: consolidate and reuse functionality  for dealing with `origin`
The axes and image base grids can differ based on the `origin` argument. We use this in the `image_base`, and should also use this in the plotting, specifically to deal with the `units` argument. Otherwise, how the stimulus is drawn does not match the units being plotted on the axes. Consolidating  this in one place ensure that we can flexibly expand this logic if we want.
