+++ stimupy/components/__init__.py
@@ -63,20 +63,8 @@ def image_base(visual_size=None, shape=None, ppd=None, rotation=0.0, origin="mea
     # Resolve resolution
     shape, visual_size, ppd = resolution.resolve(shape=shape, visual_size=visual_size, ppd=ppd)
 
-    # Set origin
-    if origin == "corner":
-        x = np.linspace(0, visual_size.width, shape.width)
-        y = np.linspace(0, visual_size.height, shape.height)
-    elif origin == "mean":
-        vrange = (visual_size.height / 2, visual_size.width / 2)
-        x = np.linspace(-vrange[1], vrange[1], shape.width)
-        y = np.linspace(-vrange[0], vrange[0], shape.height)
-    elif origin == "center":
-        vrange = (visual_size.height / 2, visual_size.width / 2)
-        x = np.linspace(-vrange[1], vrange[1], shape.width, endpoint=False)
-        y = np.linspace(-vrange[0], vrange[0], shape.height, endpoint=False)
-    else:
-        raise ValueError("origin can only be be corner, mean or center")
+    # Get single axes
+    x, y = resolution.visual_size_to_axes(visual_size=visual_size, shape=shape, origin=origin)
 
     # Linear distance image bases
     xx, yy = np.meshgrid(x, y)
