### 2023-04-06 #5321 @emdupre
Thank you both for your quick (!) and considered feedback, @alexander-pastukhov and @JonathanReardon !

To cross-link the currently raised issues:
- https://github.com/computational-psychology/stimupy/issues/74
- https://github.com/computational-psychology/stimupy/issues/73
- https://github.com/computational-psychology/stimupy/issues/72
- https://github.com/computational-psychology/stimupy/issues/71
- https://github.com/computational-psychology/stimupy/issues/70
- https://github.com/computational-psychology/stimupy/issues/69

I am able to install and import the package locally, and I can replicate the `Stimuli - Bullseyes` demo locally (with an additional `ipywidgets` dependency installation). So, it seems that the package is functional -- though with noted issues -- but the documentation, and in particular the Topic Guides, are sparsely populated.

The [JOSS reviewing guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#documentation) note that :

> There should be sufficient documentation for you, the reviewer, to understand the core functionality of the software under review. A high-level overview of this documentation should be included in a README file (or equivalent).

If you believe that there is sufficient documentation for you to understand the core functionality of the software and assess how it meets the functionality claims described in the JOSS paper, then I think we can proceed with the review. Please let me know.

@LynnSchmittwilken, please feel free to start addressing the current issues as raised by the reviewers ! You do not need to wait until a specific point to engage in this process.
