### 2023-04-06 5a41f0f
docs: better referencing, using hoverxref
