### 2023-04-06 722a977
docs(tutorial): more structured overview of `stimuli`
Closes #69
