+++ docs/getting_started/stimulus.md
@@ -10,30 +10,37 @@ kernelspec:
   name: python3
 ---
 
-# Exploring more complex stimuli
+# Exploring parameterized stimuli
 
-The geometric components in [`stimupy.components`](../reference/_api/stimupy.components)
-form the basic building blocks for all stimuli implement in `stimupy`.
-More complex stimuli can be composed using the functions that generate components.
-
-## Simultaneous Brightness Contrast (SBC)
+The geometric {py:mod}`components <stimupy.components>` are the basic building blocks
+which can be composed into all kinds of stimuli.
 
 ```{code-cell}
-:tags: [hide-cell]
 import matplotlib.pyplot as plt
 from stimupy.utils import plot_stim
 ```
 
 Included in `stimupy` is a large set of functions
-to generate more complex [`stimuli`](../reference/_api/stimupy.stimuli).
-These are generally subdivided into specifically named submodules.
-All of these can be accessed either through `stimupy.stimuli.<submodule>`,
-or by `from stimupy import <submodule>`:
+to generate known {py:mod}`stimuli <stimupy.stimuli>` (see [How stimupy is organized](../topic_guides/organization))
+These are generally subdivided into submodules
+bearing their usual name.
+```{margin}
+The submodule names are pluralized,
+so a {py:func}`bullseye <stimupy.stimuli.bullseye.bullseye>` stimulus function
+can be found in {py:func}`stimupy.stimuli.bullseyes.bullseye`
+```
+All of these can be accessed by `import stimupy.stimuli.<submodule>`
+or `from stimupy.stimuli import <submodule>`
 
+
+## Thin wrapper around component - Simultaneous Brightness Contrast (SBC)
+Firstly, a geometrically simple stimulus:
+a classic Simultaneous Brightness Contrast (SBC) display:
 ```{code-cell}
-from stimupy import sbcs
+from stimupy.stimuli import sbcs
 
-stim = sbcs.basic(visual_size=(6,8), ppd=10, target_size=(2,2))
+stim = sbcs.basic(visual_size=(6,8), ppd=10,
+                  target_size=(2,2))
 
 plot_stim(stim)
 plt.show()
@@ -41,27 +48,44 @@ plt.show()
 
 As you can see, this is not (much) more than:
 ```{code-cell}
-from stimupy.components.shapes import rectangle
+from stimupy.components import shapes
 
-component = rectangle(visual_size=(6,8), ppd=10, rectangle_size=(2,2), intensity_background=0.0, intensity_rectangle=0.5)
+component = shapes.rectangle(visual_size=(6,8), ppd=10,
+                             rectangle_size=(2,2),
+                             intensity_background=0.0, intensity_rectangle=0.5)
 
 plot_stim(component)
 plt.show()
 ```
 
-However, some of the stimulus parameters have different names in `stimuli`.
-In particular, **all** `stimuli` have the concept of a `target` region(s):
+{py:mod}`stimuli <stimupy.stimuli>`, like {py:mod}`components <stimupy.components>`,
+also have 3 overall categories of parameters:
+- image size & resolution
+- stimulus geometry
+- stimulus photometry (intensities)
+
+However, some of the stimulus parameters
+have different names in {py:mod}`stimuli <stimupy.stimuli>`.
+In particular, **all** {py:mod}`stimuli <stimupy.stimuli>`
+have the concept of a `target` region(s):
 image regions that are of some particular scientific interest in this stimulus.
 For an SBC stimulus, this would be the rectangular path.
-Thus, the `sbcs.basic` function takes a
+Thus, the {py:func}`sbcs.basic <stimupy.stimuli.sbcs.basic>` function takes a
 - `target_size`, compared to `rectangle_size`
[…]
