### 2023-04-06 a93b888
build: specify build system
Use Setuptools build-backend.
Tell PSR to build from `main`, using PyPAs Build
