### 2023-04-06 computational-psychology/stimupy#75 @JonathanReardon
empty PKG-INFO
Your PKG-INFO file is missing metadata. I would update your setup.py file (and rebuild the package) to include this information.
