### 2023-04-07 927f029
build: Versioning workflow also builds and uploads assets to release-tag
And allow manual trigger
