+++ .github/workflows/release.yml
@@ -1,0 +1,36 @@
+name: Release to PyPI
+
+on: [release, worfklow_dispatch]
+
+jobs:
+  test_publish:
+    needs: build_dist
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Fetch wheel(s) from release
+        uses: dsaltares/fetch-gh-release-asset@1.1.0
+        regex: true
+        file: '*.whl'
+        target: 'dist/'
+
+      - name: Fetch sdist(s) from release
+        uses: dsaltares/fetch-gh-release-asset@1.1.0
+        regex: true
+        file: '*.tar.gz'
+        target: 'dist/'
+
+      - name: Publish to TestPyPI
+        uses: pypa/gh-action-pypi-publish@release/v1
+        with:
+          user: __token__
+          password: ${{ secrets.TEST_PYPI_API_TOKEN }}
+          repository_url: https://test.pypi.org/legacy/
+
+      - name: Test install from TestPyPI
+        run: |
+          pip install \
+          --index-url https://test.pypi.org/simple/ \
+          --extra-index-url https://pypi.org/simple \
+          stimupy
+
