+++ .github/workflows/version.yml
@@ -16,6 +16,7 @@ jobs:
         uses: actions/checkout@v3
         with:
           fetch-depth: 0 # checkout full commit history
+          token: ${{ secrets.GHA_Token }}
 
       - name: Setup Python
         uses: actions/setup-python@v4
@@ -24,13 +25,13 @@ jobs:
 
       - name: Configure git to be able to push to repo
         env:
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          GH_TOKEN: ${{ secrets.GHA_Token }}
         run: |
-          git config user.name github-actions
-          git config user.email github-actions@github.com
+          git config user.name token
+          git config user.email token@github.com
 
       - name: Bump version, build & upload release assets using Semantic Release
         env:
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          GH_TOKEN: ${{ secrets.GHA_Token }}
         run: |
           pipx run python-semantic-release publish -v DEBUG
