### 2023-04-07 3d49f53
docs: cleanup installation instructions
Closes #71, Closes #73
