+++ docs/_toc.yml
@@ -12,11 +12,11 @@ parts:
     chapters:
     - file: topic_guides/topic_guides
     - file: topic_guides/organization
-    - file: topic_guides/sets_papers
-    - file: topic_guides/resolution
-    - file: topic_guides/dimensions
-    - file: topic_guides/gratings
-    - file: topic_guides/visual_noise
+#    - file: topic_guides/sets_papers
+#    - file: topic_guides/resolution
+#    - file: topic_guides/dimensions
+#    - file: topic_guides/gratings
+#    - file: topic_guides/visual_noise
   - caption: Reference
     chapters:
     - file: reference/api.md
