### 2023-04-07 f6a3d98
docs: Outline dependencies for demos in installation instructions
