### 2023-04-10 3d89415
v0.100.0
Automatically generated by python-semantic-release
