+++ CHANGELOG.md
@@ -1,0 +1,32 @@
+# Changelog
+
+<!--next-version-placeholder-->
+
+## v0.100.0 (2023-04-10)
+### Feature
+* **rectilinear:** Add rotation to rectilinear stimuli ([`0670d6b`](https://github.com/computational-psychology/stimupy/commit/0670d6b07b201fd0f4c0c3f8d4af6480351d2f09))
+
+### Documentation
+* Outline dependencies for demos in installation instructions ([`f6a3d98`](https://github.com/computational-psychology/stimupy/commit/f6a3d981d9e43c5567dd6237230819fc60b60403))
+* **topic guides:** Hide guides that are under construction ([`8a1146f`](https://github.com/computational-psychology/stimupy/commit/8a1146fb7968227b9add217fa0b54b2e63ccfd3c))
+* Cleanup installation instructions ([`3d49f53`](https://github.com/computational-psychology/stimupy/commit/3d49f53d7fbcb8018ce5811f1a0ab67bae75c132))
+* Update references in replication tutorial ([`2325ce3`](https://github.com/computational-psychology/stimupy/commit/2325ce37b86e8641379ba4e8412ca45251de3033))
+* **tutorial:** More structured overview of `stimuli` ([`722a977`](https://github.com/computational-psychology/stimupy/commit/722a97799997d091f3445c7e7b8a71446a7bfb45))
+* Better referencing, using hoverxref ([`5a41f0f`](https://github.com/computational-psychology/stimupy/commit/5a41f0f7e24aab0349f97faca5fed3c925997adc))
+* **tutorial:** More detail on composition, masks ([`9be4db0`](https://github.com/computational-psychology/stimupy/commit/9be4db035c37c921aefc846c89e286ad8fba56cc))
+* **tutorial:** Clean up imports ([`11053b3`](https://github.com/computational-psychology/stimupy/commit/11053b373cf98567a4a786a34d4db464efb57759))
+* **tutorial:** Fix a reference ([`506274e`](https://github.com/computational-psychology/stimupy/commit/506274ef9002cb217f466e60de45f4a7eb6d9bd1))
+* **topics:** Add page on how stimupy is organized ([`0cd3934`](https://github.com/computational-psychology/stimupy/commit/0cd39345f0647648ca37e629fad568cd124c867e))
+* **tutorial:** Move tutorial to separate TOC entry ([`cd3af04`](https://github.com/computational-psychology/stimupy/commit/cd3af04c78c25eef6f0ff4e3243365ee6620da1d))
+* **tutorial:** More verbose tutorial component ([`31c4ae6`](https://github.com/computational-psychology/stimupy/commit/31c4ae69223ba8c507155010ed509c462c4b80b1))
+* **tutorial:** Move composition to separate tutorial page ([`939082a`](https://github.com/computational-psychology/stimupy/commit/939082a5a69f6ccdc087d1e7c5ebb1ae530128c4))
+* **landing page:** Use description from README, more detailed cards ([`530dbc0`](https://github.com/computational-psychology/stimupy/commit/530dbc0d2997d4e94ae9a23f6457c05a669099c6))
+* README streamlined, in line with manuscript ([`ad8445d`](https://github.com/computational-psychology/stimupy/commit/ad8445d55955044e57ab494ddbbf5889d7b70653))
+* Detail on dev installation instructions ([`506ebf6`](https://github.com/computational-psychology/stimupy/commit/506ebf6ba60e2afc7c2e76434a03ac766bf48cff))
[…]
