### 2023-04-12 6983f90
docs(readme): fix some typos
