### 2023-04-17 #5321 @emdupre
I also wanted to cross-link another, more recently raised issue from this review : https://github.com/computational-psychology/stimupy/issues/75
