### 2023-04-17 computational-psychology/stimupy#75 @JorisVincent
Hi @JonathanReardon, could you elaborate on what metadata is missing?

When I look at the (latest, but also previous) [release](https://github.com/computational-psychology/stimupy/releases/tag/v0.100.0), specifically the [`tar.gz`](https://github.com/computational-psychology/stimupy/releases/download/v0.100.0/stimupy-0.100.0.tar.gz), the PKG-INFO in there seems fine; it has title, authors, license, classifiers, README, etc.

I'm not very familiar with the PKG-INFO process, is there some other information that should be in there? Also, stimupy uses the PEP 621 `pyproject.toml` configuration, instead of `setup.py`, maybe that changes what information ends up in PKG-INFO?
