### 2023-04-18 #5321 @JorisVincent
Hello all!

I'm one of the authors and maintainers of stimupy! :wave: Thanks for the very helpful feedback so far.

Updates from our side:

1. we've recently pushed some bugfixes, and simultaneously updated the documentation, as version `0.100.0`. Specifically, the ["Getting started" tutorial section ](https://stimupy.readthedocs.io/en/latest/getting_started/getting_started.html) should now work without errors, and the [installation instructions](https://stimupy.readthedocs.io/en/latest/getting_started/installation.html) have been cleaned up. This should close some of the issues raised, although happy to address them further if @alexander-pastukhov and/or @JonathanReardon feel like these pages are not yet clear enough. Our hope is that we this tutorial, the reviewers can get up and running to evaluate some more of the package.

2. We're also working on the topic guides, e.g. [how stimupy is organized](https://stimupy.readthedocs.io/en/latest/topic_guides/organization.html) -- a few other pages should fill in by the end of this week. The purpose of those guides, broadly, is to explain some of the concepts and design decisions of the package. The scope of that is something we could definitely use the input from the reviewers on: what are elements of the package that are not easily understood from just the tutorial plus API reference / demos? For now, we have in mind to explain:

   - [ ] image resolution (pixels to visual degrees) and how to specify it
   - [ ] an explanation of how periodic stimuli (e.g., sinewave gratings) can be specified in stimupy
   - [ ] a brief description of the various directions/distances that underlie some of the stimuli, e.g. radial vs. horizontal vs. angular gratings.

3. In the meantime, we're working on the CI/CD aspect of publishing the package on PyPI. Once this is done, installation should simplify a lot as well, thus preventing some of the earlier raised issues.

4. We're addressing some of the remaining issues individually.

I hope this gives some perspective on how we're taking in the review from our side. If there are any questions or concerns, we're very happy to address those. Thanks again for the constructive and critical review, since we really want stimupy to become a great solution to common problems in vision science, for both experts and novices, all feedback is hugely valuable to us!
