### 2023-04-18 ae98e04
docs(demos): better exception-handling, hiding stacktrace
Closes #74
