+++ docs/reference/demos/stimuli/whites.md
@@ -91,20 +91,23 @@ def show_generalized(
     target_center_offsets2=None,
     target_heights2=None,
 ):
-    stim = generalized(
-        visual_size=(height, width),
-        ppd=ppd,
-        rotation=rotation,
-        frequency=frequency,
-        intensity_bars=(int1, int2),
-        origin=origin,
-        period=period,
-        target_indices=(target_idx, target_idx2),
-        intensity_target=(intensity_target, intensity_target2),
-        target_center_offsets=(target_center_offsets, target_center_offsets2),
-        target_heights=(target_heights, target_heights2),
-    )
-    plot_stim(stim, mask=add_mask)
+    try:
+        stim = generalized(
+            visual_size=(height, width),
+            ppd=ppd,
+            rotation=rotation,
+            frequency=frequency,
+            intensity_bars=(int1, int2),
+            origin=origin,
+            period=period,
+            target_indices=(target_idx, target_idx2),
+            intensity_target=(intensity_target, intensity_target2),
+            target_center_offsets=(target_center_offsets, target_center_offsets2),
+            target_heights=(target_heights, target_heights2),
+        )
+        plot_stim(stim, mask=add_mask)
+    except Exception as e:
+        raise ValueError(f"Invalid parameter combination: {e}") from None
 
 # Set interactivity
 out = iw.interactive_output(
@@ -195,19 +198,22 @@ def show_white(
     target_heights=None,
     target_heights2=None,
 ):
-    stim = white(
-        visual_size=(height, width),
-        ppd=ppd,
-        rotation=rotation,
-        frequency=frequency,
-        intensity_bars=(int1, int2),
-        origin=origin,
-        period=period,
-        target_indices=(target_idx, target_idx2),
-        intensity_target=(intensity_target, intensity_target2),
-        target_heights=(target_heights, target_heights2),
-    )
-    plot_stim(stim, mask=add_mask)
+    try:
+        stim = white(
+            visual_size=(height, width),
+            ppd=ppd,
+            rotation=rotation,
+            frequency=frequency,
+            intensity_bars=(int1, int2),
+            origin=origin,
+            period=period,
+            target_indices=(target_idx, target_idx2),
+            intensity_target=(intensity_target, intensity_target2),
+            target_heights=(target_heights, target_heights2),
+        )
+        plot_stim(stim, mask=add_mask)
+    except Exception as e:
+        raise ValueError(f"Invalid parameter combination: {e}") from None
 
 # Set interactivity
 out = iw.interactive_output(
@@ -300,21 +306,24 @@ def show_white_two_rows(
     intensity_target2=None,
     target_heights2=None,
 ):
-    stim = white_two_rows(
-        visual_size=(height, width),
-        ppd=ppd,
-        rotation=rotation,
-        frequency=frequency,
-        intensity_bars=(int1, int2),
-        origin=origin,
-        period=period,
-        target_indices_top=target_idx,
-        target_indices_bottom=target_idx2,
-        intensity_target=(intensity_target, intensity_target2),
-        target_center_offset=target_center_offsets,
-        target_heights=(target_heights, target_heights2),
-    )
-    plot_stim(stim, mask=add_mask)
+    try:
+        stim = white_two_rows(
+            visual_size=(height, width),
+            ppd=ppd,
+            rotation=rotation,
+            frequency=frequency,
+            intensity_bars=(int1, int2),
+            origin=origin,
+            period=period,
+            target_indices_top=target_idx,
+            target_indices_bottom=target_idx2,
+            intensity_target=(intensity_target, intensity_target2),
+            target_center_offset=target_center_offsets,
[…]
