### 2023-04-18 c3b0a0e
docs(demos): make dependencies explicit
add admonition about dependencies for local running
