### 2023-04-18 c28b7be
docs: hover_xref configuration for objects
