### 2023-04-25 computational-psychology/stimupy#75 @JonathanReardon
Hello @JorisVincent, I think this had to do with the way I installed the package a few weeks ago. There were several "UNKNOWN" messages (in terminal) and a largely empty "PKG-INFO" file containing several "UNKNOWN" values. That said, since we can now successfully install the package via pip, this is no longer a problem. Feel free to close the issue. Thanks.
