### 2023-04-29 d602fc0
refactor: edges as `stimuli`-submodule
