+++ stimupy/stimuli/edges.py
@@ -1,0 +1,2 @@
+from stimupy.components.edges import *
+from stimupy.components.edges import overview
