### 2023-05-01 1679eb5
feat: util to create stimspace
