+++ stimupy/utils/utils.py
@@ -1,4 +1,5 @@
 import copy
+import itertools
 
 import numpy as np
 import scipy.special as sp
@@ -17,6 +18,7 @@ __all__ = [
     "flip_dict",
     "roll_dict",
     "strip_dict",
+    "generate_stimspace",
 ]
 
 
@@ -435,3 +437,30 @@ def strip_dict(
         if name in dct.keys():
             new_dict[name] = dct[name]
     return new_dict
+
+
+def generate_stimspace(stimulus_function, params, title_params=None):
+    if not callable(stimulus_function):
+        raise ValueError("stimulus_function needs to be a function")
+    if not isinstance(params, dict):
+        raise ValueError("params needs to be a dict with all stimulus parameters")
+    if isinstance(title_params, str):
+        title_params = [title_params, ]
+
+    keys, values = zip(*params.items())
+    permutations_dicts = [dict(zip(keys, v)) for v in itertools.product(*values)]
+
+    stimuli = {}
+    for i, p in enumerate(permutations_dicts):
+        if title_params is None:
+            key = str(i)
+        else:
+            key = ""
+            for tname in title_params:
+                if isinstance(p[tname], (float, int)):
+                    ptname = np.round(p[tname], 3)
+                    key += f"{tname}={ptname} "
+                else:
+                    key += f"{tname}={p[tname]} "
+        stimuli[key] = stimulus_function(**p, )
+    return stimuli
