### 2023-05-02 ae20b4d
refactor: separate params-permutation and stimspace-stim-creation
