### 2023-05-05 computational-psychology/stimupy#78 @alexander-pastukhov
Topic Guide: step-by-step tutorial for create and submitting stimulus for a paper
The main purpose of the package is to make it easier to deposit a stimulus, so that it can be easily replicated by others. Therefore, ot would be very useful to have a step-by-step guide on how to create a corresponding file for your own work. The stimulus itself could be very simple, it is more about the minimal file structure, necessary functions, documentation, etc.

In addition, this guide should contain information on how best to distribute it. Is there a way to include my own stimulus into the package apart from a pull request? Is it even an intention of the authors to turn the package into the stimulus repository (in this case CRAN-like submission and website with a gallery would be very useful)? Should one use the package but keep the stimulus as part of separate open methods files? Guidelines on this would be very important to build stimuli community.
