### 2023-05-05 computational-psychology/stimupy#79 @alexander-pastukhov
Topic Guide: using stimuli with PsychoPy, PTB, etc.
It would be very useful to have a brief guide with simple stimuli on basic integration with experimental code, just to illustrate best/simplest/recommended ways of doing this.
