### 2023-05-05 #5321 @alexander-pastukhov
@emdupre Sorry, somehow lost the track of time. The authors address my initial concerns but I've raised two more issues regarding documentation. I feel that having topic guides on 1) how to use stimuli generated by the package in other software, and 2) how to create and deposit your own stimulus for later replication, are critical as they relate to the core use of the package (create and use stimulus for my research, then tell others how to replicate it).
