### 2023-05-05 f9a5a4d
fix(mondrians): clearer error when neither nrows, ncols, depths or intensities are specified
