### 2023-05-05 bbab735
refactor(export): consistent naming of exporting funcs
and rename `arr_to_checksum`
