+++ stimupy/utils/export.py
@@ -5,15 +5,49 @@ import numpy as np
 from PIL import Image
 
 __all__ = [
-    "arrs_to_checksum",
+    "array_to_checksum",
+    "array_to_image",
+    "arrays_to_checksum",
     "to_json",
-    "write_array_to_image",
 ]
 
 
-def arrs_to_checksum(stim, keys=["img", "mask"]):
+def array_to_checksum(arr):
     """
     Hash (md5) values, and save only the hex
+
+    Parameters
+    ----------
+    arr : np.ndarray
+
+    Returns
+    ----------
+    hex
+    """
+    return md5(np.ascontiguousarray(arr.round(8))).hexdigest()
+
+
+def array_to_image(arr, filename):
+    """
+    Save a 2D numpy array as a grayscale image file.
+
+    Parameters
+    ----------
+    arr : np.ndarray
+        The data to be stored in the image. Values will be cropped to [0,255].
+    filename : str
+        full path to the file to be creaated.
+    """
+    if Image:
+        imsize = arr.shape
+        im = Image.new("L", (imsize[1], imsize[0]))
+        im.putdata(arr.flatten())
+        im.save(filename)
+
+
+def arrays_to_checksum(stim, keys=["img", "mask"]):
+    """
+    Hash (md5) values of arrays specified in keys, and save only the hex
 
     Parameters
     ----------
@@ -33,7 +67,7 @@ def arrs_to_checksum(stim, keys=["img", "mask"]):
         ]
 
     for key in keys:
-        stim[key] = md5(np.ascontiguousarray(stim[key].round(8))).hexdigest()
+        stim[key] = array_to_checksum(stim[key])
 
     return stim
 
@@ -53,22 +87,3 @@ def to_json(stim, filename):
     # stimulus-dict(s) as (pretty) JSON
     with open(filename, "w", encoding="utf-8") as f:
         json.dump(stim, f, ensure_ascii=False, indent=4)
-
-
-def write_array_to_image(filename, arr):
-    """
-    Save a 2D numpy array as a grayscale image file.
-
-    Parameters
-    ----------
-    filename : str
-        full path to the file to be creaated.
-    arr : np.ndarray
-        2D numpy array
-        The data to be stored in the image. Values will be cropped to [0,255].
-    """
-    if Image:
-        imsize = arr.shape
-        im = Image.new("L", (imsize[1], imsize[0]))
-        im.putdata(arr.flatten())
-        im.save(filename)
