+++ stimupy/stimuli/edges.py
@@ -1,2 +1,2 @@
-from stimupy.components.edges import *
-from stimupy.components.edges import overview
+from stimupy.components.edges import *  # noqa: F401
+from stimupy.components.edges import overview  # noqa: F401
