+++ stimupy/utils/export.py
@@ -1,14 +1,22 @@
 import json
 from hashlib import md5
 
+import copy
 import numpy as np
 from PIL import Image
+from scipy.io import savemat
+import pickle
 
 __all__ = [
     "array_to_checksum",
     "array_to_image",
+    "array_to_npy",
+    "array_to_mat",
+    "array_to_pickle",
     "arrays_to_checksum",
     "to_json",
+    "to_mat",
+    "to_pickle",
 ]
 
 
@@ -19,6 +27,7 @@ def array_to_checksum(arr):
     Parameters
     ----------
     arr : np.ndarray
+        Array to be hashed.
 
     Returns
     ----------
@@ -27,22 +36,117 @@ def array_to_checksum(arr):
     return md5(np.ascontiguousarray(arr.round(8))).hexdigest()
 
 
-def array_to_image(arr, filename):
+def array_to_image(arr, filename, norm=True):
     """
     Save a 2D numpy array as a grayscale image file.
 
     Parameters
     ----------
     arr : np.ndarray
-        The data to be stored in the image. Values will be cropped to [0,255].
+        Array to be saved. Values will be cropped to [0,255].
+    filename : str
+        full path to the file to be created.
+    norm : bool
+        if True (default), multiply array by 255
+    """
+    if filename[-4:] != ".png" and filename[-4:] != ".jpg":
+        filename += ".png"
+
+    if isinstance(arr, (np.ndarray, list)):
+        arr = np.array(arr)
+
+        if norm:
+            arr = arr * 255
+
+        if Image:
+            imsize = arr.shape
+            im = Image.new("L", (imsize[1], imsize[0]))
+            im.putdata(arr.flatten())
+            im.save(filename)
+    else:
+        raise ValueError("arr should be a np.ndarray")
+
+
+def array_to_npy(arr, filename):
+    """
+    Save a numpy array to npy-file.
+
+    Parameters
+    ----------
+    arr : np.ndarray
+        Array to be saved.
     filename : str
         full path to the file to be creaated.
     """
-    if Image:
-        imsize = arr.shape
-        im = Image.new("L", (imsize[1], imsize[0]))
-        im.putdata(arr.flatten())
-        im.save(filename)
+    if isinstance(arr, (np.ndarray, list)):
+        np.save(filename, arr)
+    else:
+        raise ValueError("arr should be a np.ndarray")
+
+
+def array_to_mat(arr, filename):
+    """
+    Save a numpy array to a mat-file.
+
+    Parameters
+    ----------
+    arr : np.ndarray
+        Array to be saved.
+    filename : str
+        full path to the file to be creaated.
+    """
+    if filename[-4:] != ".mat":
+        filename += ".mat"
+
+    if isinstance(arr, (np.ndarray, list)):
+        savemat(filename, {"arr": arr})
+    else:
+        raise ValueError("arr should be a np.ndarray")
+
+
+def array_to_pickle(arr, filename):
+    """
+    Save a numpy array to a pickle-file.
+
+    Parameters
+    ----------
+    arr : np.ndarray
+        Array to be saved.
+    filename : str
+        full path to the file to be creaated.
+    """
+    if filename[-7:] != ".pickle":
+        filename += ".pickle"
+
+    if isinstance(arr, (np.ndarray, list)):
+        with open(filename, 'wb') as handle:
+            pickle.dump({"arr": arr}, handle, protocol=pickle.HIGHEST_PROTOCOL)
+    else:
+        raise ValueError("arr should be a np.ndarray")
+
+
+def array_to_json(arr, filename):
+    """
+    Save a numpy array to a (pretty) JSON.
+
+    Parameters
+    ----------
+    arr : np.ndarray
[…]
