### 2023-05-09 #5321 @emdupre
Thank you for the update, @alexander-pastukhov  !

I'm cross-linking the two newly raised issues for clarity:
- https://github.com/computational-psychology/stimupy/issues/78
- https://github.com/computational-psychology/stimupy/issues/79
