+++ docs/_toc.yml
@@ -10,10 +10,9 @@ parts:
     - file: getting_started/getting_started
   - caption: Topic guides
     chapters:
-    - file: topic_guides/topic_guides
     - file: topic_guides/organization
 #    - file: topic_guides/sets_papers
-#    - file: topic_guides/resolution
+    - file: topic_guides/resolution
 #    - file: topic_guides/dimensions
 #    - file: topic_guides/gratings
 #    - file: topic_guides/visual_noise
