### 2023-05-10 17ff040
docs(contribute): instructions and examples for contributing paper set
