+++ docs/contributing/sets.md
@@ -1,13 +1,133 @@
 # Contribrute a stimulus (set)
+`stimupy` contains both a wide range of generalized [stimulus-functions](stimupy.stimuli)
+that can generate many parameterizated stimulus images,
+and several complete sets of stimuli
+that all come from the same {py:mod}`source <stimupy.papers>`.
+Development, organization and maintanance of these sets can be arduous,
+and we are committed to providing a high-quality and well-tested suite of functions.
+Thus, `stimupy` is a curated and maintained _library_ of stimuli
+(stimulus-functions, to be precise),
+not a _repository_ or platform for uploading stimuli.
+We welcome contributions to this library of stimuli,
+and stimulus sets, from the literature.
 
 ## Add a stimulus set
-- Create a new file in `stimupy/papers`
-- Add only stimuli that use `stimupy` functionality
+If you wish to contribute a complete set of stimuli from a single paper,
+please follow these instructions for organizing and formatting the code.
+
+A {py:mod}`paper <stimupy.papers>` should have the following features:
+
+- Each paper is identified using a `paper_key`: a cite-key like indicator,
+  usually some combination of author names/initials and publication date.
+  - For {py:mod}`modelfest <stimupy.papers.modelfest>` we made an exception.
+- All stimuli belonging a `paper` should be in a single python _module_, i.e., `.py` code-file;
+  this file is found under `stimupy/papers/<paper_key>.py`
+- Only stimuli that `stimupy` functionality should be included in the `<paper_key>.py`
+  - We have made a select few exceptions to this rule
 - Each stimulus should be its own function
 - Each function should only take a `ppd` argument;
   it should replicate the exact size and geometry of the stimulus,
   allowing only the resolution (i.e., pixels per degree) to be changed.
-- Also create a corresponding `tests/papers_my_paper.py` file
+  This argument should have an actual default value,
+  which corresponds to the `ppd` used in the original project.
+- Thus, different parameterizations of the "same" stimulus
+  (e.g. different geometries, polarities, etc.) should be separate functions.
+- As a result, each function should run without any arguments, e.g.
+  ```{code-block} python
+  stim = stimupy.papers.RHS2007.WE_thick()
+  ```
+- The output stim-`dict`s must contain at least the `img`,
+  but should also contain relevant stimulus parameters.
+- The output stim-`dict`s may contain additional information (keys),
+  such as experimental results,
+  if they do not take up too much space.
+- Each function should have a [NumPy-style docstring](https://numpydoc.readthedocs.io/en/latest/format.html#docstring-standard),
+  which at least documents the `ppd` argument and the returned dict as follows:
+  ```{code-block} python
+    """ <short summary of stimulus function>
+
+    Parameters
+    ----------
+    ppd : int
+        Resolution of stimulus in pixels per degree. (default: 32)
+
+    Returns
+    -------
+    dict of str
+        dict with the stimulus (key: "img")
+        and additional keys containing stimulus parameters
+    """
+  ```
+  Ideally it also provides detailed information of which exact stimulus it is,
+  e.g., through reference to a specific figure in the original paper.
+- The paper-module should also contain an `__all__` attribute,
+  which contains a list of all function names of the stimuli in the paper.
+  Anything in here is considered a stimulus, anything not in here will not be available
+  outside the paper module itself.
+- The paper-module should have a NumPy style docstring as well,
+  documenting at least the full bibliographical reference for the source of the stimuli.
+
+### An example
+```{tip}
+This file is included as `stimupy/papers/example.py`
+and can be `import`ed as `stimupy.papers.example`
+```
+
[…]
