### 2023-05-10 a733b07
docs(howto): display stimuli (in experiment)
Closes #79
