+++ docs/howtos/share.md
@@ -1,0 +1,144 @@
+# Share & distribute stimuli
+A core principle of `stimupy` is sharing and reusing stimuli.
+One part of this is the large number of existing stimuli already implement,
+such that you can generate various parameterizations of these
+without having to write any custom code.
+The other part is that, hopefully, you will share the `stimupy`-stimuli
+that you use in your work.
+
+
+## What to share?
+To accurately reproduce a stimulus,
+a complete description of the parameters (size, geometry, radiometry)
+should be enough.
+However, in developing `stimupy` it has been our experience that
+such descriptions are often incomplete and parameters can be easily overlooked.
+
+Another option is to share the exact digital image (i.e., the matrix of pixel values)
+of the stimulus.
+While this allows that exact digital image to be reused,
+any modifications may be hard to implement --
+especially changing the [resolution](../topic_guides/resolution.md).
+Moreover, sets of digital images can take up quite some storage space,
+and are not ideal for easy and quick transfer.
+
+Instead, we believe that the _code_ underlying stimulus creation should be shared,
+alongside the parameters used to call that code.
+The combination of the same exact code plus the same parameters should *always*
+produce the exact same digital image.
+Thus, you can share the code and parameters,
+and a recipient can reproduce your exact stimulus.
+
+This is why we distribute `stimupy`:
+it contains lots of (generalized) code to generate lots of (parameterizable) stimuli.
+When you use this to generate specific stimuli in your own work
+there are several ways to share the require code and parameters,
+depending on whether you "just" used specific parameterization of a function in `stimupy`,
+or added some custom code (e.g., composition).
+
+````{important}
+In general, the following should be communicated / shared:
+
+0. Platform and environment information
+  - Operating System (and version)
+  - Python version
+  - versions of all other requisite packages:
+    A full list of installed Python packages and their version numbers can be generated
+    using (from outside Python):
+    ```{code-block} bash
+    pip freeze
+    ```
+
+1. exact `stimupy` version
+   This version can be found using (from within Python):
+   ```{code-block} python
+   >>> import stimupy
+   >>> stimupy.__version__
+   '1.0.0'
+   >>>
+   ```
+2. parameters for stimupy-stimuli
+3. any custom code, e.g. for composition
+````
+
+## Existing stimulus, specific parameterization
+
+If your work uses a stimulus as it is implemented in `stimupy`,
+e.g., some {py:func}`gabor <stimupy.stimuli.gabors.gabor>`
+or some parameterization of the {py:func}`Todorovic illusion <stimupy.stimuli.todorovics.cross>`,
+without any additional customization / custom code (e.g., composition),
+you only need to share the parameter-values for this stimulus.
+All the code necessary to reproduce the stimulus from those values
+is included in `stimupy`.
+Thus, you'll need to:
+
+- indicate which exact version of `stimupy` you used.
+  A recipient can then install this version of `stimupy`
+  and reproduce the exact stimulus from the provide parameter values.
+
+- provide all parameter-values for this stimulus, ideally by [exporting them](export.md).
+  For just sharing paramter-values, we recommend sharing the stimulus-`dict`
+  without the `img` and `masks` to a [JSON file](./export.md).
+  This because JSON files are both human-readable -- so that a non-stimupy / non-Python
+  recipient can still evaluate the parameter values -- and safely machine-readable
+  -- so that a Python / stimupy user can easily load and recreate the stimulus.
+
+%% SOMETHING ABOUT HASHING
+
+## Custom stimulus
+If your stimulus is not just a parameterization of an existing `stimupy` function
+and contains some custom code (e.g., composition),
+you will also need to distribute your custom code.
+This can be done in several ways.
+
+If you are already sharing all the code
[…]
