### 2023-05-10 4d5b507
docs(guides): waves & gratings
