+++ docs/topic_guides/waves.md
@@ -1,0 +1,299 @@
+---
+jupytext:
+  formats: md:myst
+  text_representation:
+    extension: .md
+    format_name: myst
+kernelspec:
+  display_name: Python 3
+  language: python
+  name: python3
+---
+
+# Periodic stimuli (waves, gratings, grids,...)
+```{tip}
+This guide explains the principles behind, behavior of, and relations between
+the parameters of periodic stimuli in `stimupy`.
+For a more hands-on demonstration of how those paramaters affect a specific stimulus,
+see, e.g., [](../reference/demos/stimuli/waves.md)
+```
+
+Many commonly used visual stimuli are some sort of periodic stimulus:
+sinewaves gratings are the most obvious example,
+but also checkerboards, grids, sets of rings, and pinwheels can be considered periodic.
+`stimupy` includes many of these as stimuli and constructs them as periodic,
+which has some particular advantages.
+The simplest variants can be found in {py:mod}`stimupy.stimuli.waves`
+
+```{code-cell}
+from stimupy.stimuli import waves
+
+from stimupy.utils import plot_stimuli
+```
+
+## Interdependence of parameters
+Periodic stimuli (e.g., `waves`)
+consist of repetitions of some _elements_ or _phases_.
+Periodic stimuli (e.g., `waves`)
+consist of repetitions of some _elements_ or _phases_.
+In a sinewave grating, these are the light and dark phases.
+in a squarewave grating, those same phases visually appears as white and black `bars`:
+```{code-cell}
+sinewave = waves.sine_linear(visual_size=(10, 10), ppd=10, n_bars=5)
+squarewave = waves.square_linear(visual_size=(10, 10), ppd=10, n_bars=5)
+
+plot_stimuli({'Sine': sinewave, 'Square': squarewave})
+```
+
+There is regularity to periodic stimuli (what makes the periodic),
+namely that there are repeated _cycles_ (here of 2 phases/elements per cycle),
+where each cycle is identical to each other cycle.
+This regularity can be expressed in several different ways,
+with several different parameters 
+-- and these parameters have some interdependence on eachother.
+
+```{tip}
+In the examples in this section,
+the code for the `sinewave` will remain the same across examples,
+the code for the `squarewave` will demonstrate the different parameterization
+(keeping the output images the same)
+```
+
+### Frequency $\sim$ number of elements
+The number of elements can also be expressed a _spatial frequency_:
+here we have $5$ bars, which corresponds to $5/2 = 2.5$ _cycles_ of light-dark.
+$2.5$ cycles, in $10$ degrees of visual angle,
+is $2.5/10 = .25$ cycles per degree  of visual angle (_cpd_).
+
+In some cases, you may want to conceptualize your stimulus in terms of this `frequency`,
+in other cases you may want to conceptualize your stimulus in terms of the `n_bars`.
+`stimupy` allows you to do either:
+```{code-cell}
+sinewave = waves.sine_linear(visual_size=(10, 10), ppd=10, n_bars=5)
+squarewave = waves.square_linear(visual_size=(10, 10), ppd=10, frequency=.25)
+
+plot_stimuli({'Sine': sinewave, 'Square': squarewave})
+```
+
+### Phase width vs. visual size
+To specify a periodic stimulus like these waves, a third parameter at play is
+the `phase_width`, i.e., the size of each element (in this case: `bar_width`).
+So far, we have specified the overall `visual_size` of the stimulus (e.g., `(10, 10)`).
+To draw `n_bars=5` ($=$ `frequency=.25`) in $10$ degrees,
+each bar (/ phase) has `bar_width` of $2$ degrees of visual of angle.
+
[…]
