### 2023-05-10 7bc02ec
docs(guide): axes, orientations, distance_metrics
