+++ docs/_toc.yml
@@ -13,8 +13,8 @@ parts:
     chapters:
     - file: topic_guides/organization
     - file: topic_guides/resolution
+    - file: topic_guides/dimensions
     - file: topic_guides/waves
-#    - file: topic_guides/dimensions
 #    - file: topic_guides/visual_noise
   - caption: Reference
     chapters:
