+++ .github/workflows/release.yml
@@ -11,15 +11,17 @@ jobs:
     steps:
       - name: Fetch wheel(s) from release
         uses: dsaltares/fetch-gh-release-asset@1.1.0
-        regex: true
-        file: '*.whl'
-        target: 'dist/'
+        with:
+          regex: true
+          file: 'stimupy-.*\\.whl'
+          target: 'dist/'
 
       - name: Fetch sdist(s) from release
         uses: dsaltares/fetch-gh-release-asset@1.1.0
-        regex: true
-        file: '*.tar.gz'
-        target: 'dist/'
+        with:
+          regex: true
+          file: 'stimupy-.*\\.tar.gz'
+          target: 'dist/'
 
       - name: Publish to TestPyPI
         uses: pypa/gh-action-pypi-publish@release/v1
