+++ stimupy/papers/example.py
@@ -18,13 +18,14 @@ reference information here
 from stimupy.components import combine_masks, draw_regions, shapes
 
 # Define original size resolution parameters
-VISUAL_SIZE = (10,12)
+VISUAL_SIZE = (10, 12)
 PPD = 10
 
 __all__ = [
     "my_bullseye",
     "my_inverse_bullseye",
 ]
+
 
 # Helper function:
 def bullseye_geometry(ppd=PPD):
@@ -40,27 +41,28 @@ def bullseye_geometry(ppd=PPD):
 
     """
     # Create center (target) disc:
-    disc = shapes.disc(visual_size=VISUAL_SIZE, ppd=ppd,
-                    radius=2,
-                    intensity_disc=.5, intensity_background=.5)
+    disc = shapes.disc(
+        visual_size=VISUAL_SIZE, ppd=ppd, radius=2, intensity_disc=0.5, intensity_background=0.5
+    )
 
     # Create first ring, white:
-    ring_1 = shapes.ring(visual_size=VISUAL_SIZE, ppd=ppd,
-                        radii=(2, 3),
-                        intensity_ring=1, intensity_background=.5)
+    ring_1 = shapes.ring(
+        visual_size=VISUAL_SIZE, ppd=ppd, radii=(2, 3), intensity_ring=1, intensity_background=0.5
+    )
 
     # Create second ring, black:
-    ring_2 = shapes.ring(visual_size=VISUAL_SIZE, ppd=ppd,
-                        radii=(3, 4),
-                        intensity_ring=0, intensity_background=.5)
+    ring_2 = shapes.ring(
+        visual_size=VISUAL_SIZE, ppd=ppd, radii=(3, 4), intensity_ring=0, intensity_background=0.5
+    )
 
     bullseye_mask = combine_masks(disc["ring_mask"], ring_1["ring_mask"], ring_2["ring_mask"])
 
     return bullseye_mask
 
+
 # New stimulus function:
 def my_bullseye(ppd=PPD):
-    """ My bullseye stimulus: grey on white on black
+    """My bullseye stimulus: grey on white on black
 
     Parameters
     ----------
@@ -78,24 +80,27 @@ def my_bullseye(ppd=PPD):
     # Call geometry helper function
     bullseye_mask = bullseye_geometry(ppd=ppd)
 
-    bullseye_img = draw_regions(mask=bullseye_mask, intensities=[0.5, 1, 0], intensity_background=0.5)
+    bullseye_img = draw_regions(
+        mask=bullseye_mask, intensities=[0.5, 1, 0], intensity_background=0.5
+    )
 
     # Package into stim-dict, adding parameter information
     stim = {
-    "img": bullseye_img,
-    "visual_size": VISUAL_SIZE,
-    "ppd": ppd,
-    "radii": (2, 3, 4),
-    "intensities": (0.5, 1, 0),
-    "intensity_background": 0.5,
+        "img": bullseye_img,
+        "visual_size": VISUAL_SIZE,
+        "ppd": ppd,
+        "radii": (2, 3, 4),
+        "intensities": (0.5, 1, 0),
+        "intensity_background": 0.5,
     }
 
     # Output
     return stim
 
+
 # Second stimulus function:
 def my_inverse_bullseye(ppd=PPD):
-    """ My other bullseye stimulus: grey on black on white
+    """My other bullseye stimulus: grey on black on white
 
 
     Parameters
@@ -114,17 +119,19 @@ def my_inverse_bullseye(ppd=PPD):
     # Call geometry helper function
     bullseye_mask = bullseye_geometry(ppd=ppd)
 
-    bullseye_img = draw_regions(mask=bullseye_mask, intensities=[0.5, 0, 1], intensity_background=0.5)
[…]
