### 2023-05-15 0b5b028
feat: added stimupy logo module, and function
