+++ stimupy/__init__.py
@@ -2,5 +2,6 @@ __version__ = "0.100.0"
 
 from stimupy import components, noises, stimuli, utils
 from stimupy.stimuli import *
+from stimupy.logos import *
 
 __all__ = ["components", "noises", "utils", *stimuli.__all__]
