+++ stimupy/logos.py
@@ -1,6 +1,7 @@
 import stimupy
 import copy
 import numpy as np
+
 np.random.seed(1)
 
 __all__ = ["logo"]
@@ -8,9 +9,9 @@ __all__ = ["logo"]
 
 def logo(
     ppd=128,
-    ):
+):
     """Generate stimupy logo
-    
+
     Parameters
     ----------
     ppd : Number (default: 128)
@@ -27,50 +28,61 @@ def logo(
     res = {"visual_size": 11, "ppd": ppd}
     radius = 5
 
-    segments = stimupy.waves.square_angular(**res, n_segments=8, rotation=-(360/16))
+    segments = stimupy.waves.square_angular(**res, n_segments=8, rotation=-(360 / 16))
     circle = stimupy.components.shapes.circle(**res, radius=radius, intensity_background=0)
     pie = copy.deepcopy(segments)
 
     pie.update(
         img=np.where(circle["circle_mask"], segments["img"], 0.5),
         mask=np.where(circle["circle_mask"], segments["grating_mask"], 0),
-        radius=circle["radius"]
+        radius=circle["radius"],
     )
-    
+
     # Create stimuli for pies
-    grating = stimupy.stimuli.waves.sine_linear(**res, frequency=1, rotation=(360/8))
-    dotted = stimupy.stimuli.sbcs.dotted(**res, n_dots=10, dot_radius=0.25, distance=0.1, target_shape=2)
-    dotted = stimupy.utils.roll_dict(dotted, shift=-dotted["shape"][0]/4, axes=0)
+    grating = stimupy.stimuli.waves.sine_linear(**res, frequency=1, rotation=(360 / 8))
+    dotted = stimupy.stimuli.sbcs.dotted(
+        **res, n_dots=10, dot_radius=0.25, distance=0.1, target_shape=2
+    )
+    dotted = stimupy.utils.roll_dict(dotted, shift=-dotted["shape"][0] / 4, axes=0)
     grid = stimupy.stimuli.hermanns.grid(**res, element_size=(0.5, 0.5, 0.07))
     whites = stimupy.stimuli.whites.radial(**res, target_indices=(7, 14), frequency=2)
-    plaid = stimupy.stimuli.plaids.sine_waves(grating_parameters1={**res, "frequency": 2},
-                                              grating_parameters2={**res, "frequency": 1, "rotation": 45})
-    mondrian = stimupy.mondrians.corrugated_mondrian(**res, intensities=np.random.rand(10, 10),
-                                                     depths=(0,0,0,0,0,0,0,1,-1,0))
+    plaid = stimupy.stimuli.plaids.sine_waves(
+        grating_parameters1={**res, "frequency": 2},
+        grating_parameters2={**res, "frequency": 1, "rotation": 45},
+    )
+    mondrian = stimupy.mondrians.corrugated_mondrian(
+        **res, intensities=np.random.rand(10, 10), depths=(0, 0, 0, 0, 0, 0, 0, 1, -1, 0)
+    )
     noise = stimupy.noises.naturals.pink(**res)
-    
+
     # Fill segments
-    pie["img"] = np.where(pie["mask"]==1, .5, pie["img"])
-    pie["img"] = np.where(pie["mask"]==2, grating["img"], pie["img"])
-    pie["img"] = np.where(pie["mask"]==3, dotted["img"], pie["img"])
-    pie["img"] = np.where(pie["mask"]==4, grid["img"], pie["img"])
-    pie["img"] = np.where(pie["mask"]==5, whites["img"], pie["img"])
[…]
