### 2023-05-15 4c30a3b
fix: added missing demo scripts for edges
