### 2023-05-15 10b7745
fix: line_width in ellipse
