### 2023-05-15 a14648e
feat: qr-code with logo embedded
