### 2023-05-15 4dfabb0
docs: add logo to README
