+++ stimupy/logos.py
@@ -102,35 +102,51 @@ def qrcode(ppd=128):
     im_logo.putdata(arr.flatten())
 
     qr = qrcode.QRCode(error_correction=qrcode.constants.ERROR_CORRECT_H)
-    qr.add_data('https://stimupy.readthedocs.io/')
-    img = qr.make_image(image_factory=StyledPilImage,
-                module_drawer=qrcode.image.styles.moduledrawers.CircleModuleDrawer(),
-                #eye_drawer=qrcode.image.styles.moduledrawers.CircleModuleDrawer(),
-                embeded_image=im_logo)
+    qr.add_data("https://stimupy.readthedocs.io/")
+    img = qr.make_image(
+        image_factory=StyledPilImage,
+        module_drawer=qrcode.image.styles.moduledrawers.CircleModuleDrawer(),
+        # eye_drawer=qrcode.image.styles.moduledrawers.CircleModuleDrawer(),
+        embeded_image=im_logo,
+    )
 
     ## Make "eyes" circular
-    eye_size = 70 #default
-    quiet_zone = 40 #default
+    eye_size = 70  # default
+    quiet_zone = 40  # default
 
-    eyes_mask = Image.new('L', img.size, 0)
+    eyes_mask = Image.new("L", img.size, 0)
 
-    eye_border = quiet_zone+eye_size
+    eye_border = quiet_zone + eye_size
 
     draw = ImageDraw.Draw(eyes_mask)
     draw.rectangle((quiet_zone, quiet_zone, eye_border, eye_border), fill=255)
-    draw.rectangle((img.size[0]-eye_border, quiet_zone, img.size[0]-quiet_zone, eye_border), fill=255)
-    draw.rectangle((quiet_zone, img.size[0]-eye_border, eye_border, img.size[0]-quiet_zone), fill=255)
+    draw.rectangle(
+        (img.size[0] - eye_border, quiet_zone, img.size[0] - quiet_zone, eye_border), fill=255
+    )
+    draw.rectangle(
+        (quiet_zone, img.size[0] - eye_border, eye_border, img.size[0] - quiet_zone), fill=255
+    )
 
     draw.rectangle((58, 58, 92, 92), fill=0)
-    draw.rectangle((img.size[0]-58, 92, img.size[0]-92, 58), fill=0)
-    draw.rectangle((55, img.size[0]-92, 92, img.size[0]-58), fill=0)
+    draw.rectangle((img.size[0] - 58, 92, img.size[0] - 92, 58), fill=0)
+    draw.rectangle((55, img.size[0] - 92, 92, img.size[0] - 58), fill=0)
 
-    round_eyes = Image.new('L', img.size, 255)
+    round_eyes = Image.new("L", img.size, 255)
 
     draw = ImageDraw.Draw(round_eyes)
     draw.ellipse((quiet_zone, quiet_zone, eye_border, eye_border), fill=None, outline=0, width=8)
-    draw.ellipse((img.size[0]-eye_border, quiet_zone, img.size[0]-quiet_zone, eye_border), fill=None, outline=0, width=8)
-    draw.ellipse((quiet_zone, img.size[0]-eye_border, eye_border, img.size[0]-quiet_zone), fill=None, outline=0, width=8)
+    draw.ellipse(
+        (img.size[0] - eye_border, quiet_zone, img.size[0] - quiet_zone, eye_border),
+        fill=None,
+        outline=0,
+        width=8,
+    )
+    draw.ellipse(
+        (quiet_zone, img.size[0] - eye_border, eye_border, img.size[0] - quiet_zone),
+        fill=None,
+        outline=0,
+        width=8,
+    )
 
     final = Image.composite(round_eyes, img, eyes_mask)
 
