+++ CHANGELOG.md
@@ -2,6 +2,43 @@
 
 <!--next-version-placeholder-->
 
+## v0.101.0 (2023-05-15)
+### Feature
+* Qr-code with logo embedded ([`a14648e`](https://github.com/computational-psychology/stimupy/commit/a14648eb05f8469e9780aedf2db6d90f3c551722))
+* Added stimupy logo module, and function ([`0b5b028`](https://github.com/computational-psychology/stimupy/commit/0b5b0281fc444d8a58fce37c9144ce2a882e3f34))
+* **export:** To image, npy, mat, pickle, json ([`3c09c5a`](https://github.com/computational-psychology/stimupy/commit/3c09c5a39f4fabed91f67096df22f8bb920a8002))
+* Util to create stimspace ([`1679eb5`](https://github.com/computational-psychology/stimupy/commit/1679eb5def54a97f826c3cb7cbbc1a982f194f18))
+
+### Fix
+* Line_width in ellipse ([`10b7745`](https://github.com/computational-psychology/stimupy/commit/10b77458186e331c721ee3a1d488ab3d32251688))
+* Added missing demo scripts for edges ([`4c30a3b`](https://github.com/computational-psychology/stimupy/commit/4c30a3bc4a8d2648e9c4bb2e3ab9055916c82762))
+* Corrected n_cells in dungeon-resolving ([`cd0fc0c`](https://github.com/computational-psychology/stimupy/commit/cd0fc0ca09a85bb829ca23bfe72171c05c66d6db))
+* Aligned functions args and keys in outputs dicts ([`107edb4`](https://github.com/computational-psychology/stimupy/commit/107edb40a5030274b10de4a962b19939d26818ab))
+* Aligned functions args and keys in output dicts ([`a3941e7`](https://github.com/computational-psychology/stimupy/commit/a3941e76c43554df8049215ecf439392bc11df1c))
+* Aligned function-args and keys in output dicts ([`8ab889a`](https://github.com/computational-psychology/stimupy/commit/8ab889a30a933f4f915f9670250265b5977bacca))
+* Radial and rectilinear waves now correctly  resolve n_phases ([`ecce0b9`](https://github.com/computational-psychology/stimupy/commit/ecce0b9c057ea4eda7d9fc4e3c3ea9b2dff0d008))
+* **mondrians:** Clearer error when neither nrows, ncols, depths or intensities are specified ([`f9a5a4d`](https://github.com/computational-psychology/stimupy/commit/f9a5a4d04048040bab367c2b6bdd62307977f657))
+* Bug in calculating plot-extent from stim-keys ([`240a34a`](https://github.com/computational-psychology/stimupy/commit/240a34a66fa733295bfbc62a154b15a434e9728e))
+
+### Documentation
+* Add logo to README ([`4dfabb0`](https://github.com/computational-psychology/stimupy/commit/4dfabb0719ea32c41425e8d8b2cc1f95097b9094))
+* Typos ([`ce4683d`](https://github.com/computational-psychology/stimupy/commit/ce4683d3f40f0a54530874ad423b5d7140443f75))
+* Fixed incorrect link in README ([`c2e6575`](https://github.com/computational-psychology/stimupy/commit/c2e65751a79d6b1b13790a49ac7bf700ffa2e6dc))
+* Updated documentation - mostly language ([`d8f4211`](https://github.com/computational-psychology/stimupy/commit/d8f4211e108f1161fda3cb2c43450caa5822d786))
+* **guide:** Axes, orientations, distance_metrics ([`7bc02ec`](https://github.com/computational-psychology/stimupy/commit/7bc02ecb7fac8bfd398ff9fe2e91542a534c172e))
[…]
