### 2023-05-15 fc8deb4
docs: minor fix landing page
