### 2023-05-15 computational-psychology/stimupy#78 @JorisVincent
In our latest update to the documentation, we have elaborated on how `stimupy`-stimuli can be used and shared, and how to contribute to `stimupy`. We had decided against being an open repository where everyone can upload a stimulus, instead going for a more curated approach via PRs. We recognize that in the long run, we may want to change this strategy, and happy to discuss alternatives.
