### 2023-05-16 fc19d2a
build: fix regex for release GHA
