+++ .github/workflows/release.yml
@@ -13,14 +13,14 @@ jobs:
         uses: dsaltares/fetch-gh-release-asset@1.1.0
         with:
           regex: true
-          file: 'stimupy-.*\\.whl'
+          file: 'stimupy-.[0-9]*\.[0-9]*\.[0-9]*\.whl'
           target: 'dist/'
 
       - name: Fetch sdist(s) from release
         uses: dsaltares/fetch-gh-release-asset@1.1.0
         with:
           regex: true
-          file: 'stimupy-.*\\.tar.gz'
+          file: 'stimupy-.[0-9]*\.[0-9]*\.[0-9]*\.tar.gz'
           target: 'dist/'
 
       - name: Publish to TestPyPI
