### 2023-05-18 #5321 @emdupre
Thank you for confirming, @JorisVincent ! It's great to see these issues addressed.

@JonathanReardon and @alexander-pastukhov , could you please revisit your reviewer checklists  and confirm if you feel that `stimupy` now meets JOSS publication standards, given these revisions ?

If you have any remaining concerns, of course, please let me know !
