+++ pyproject.toml
@@ -22,7 +22,7 @@ maintainers = [
   {name = "Lynn Schmittwilken", email = "l.schmittwilken@tu-berlin.de"},
   {name = "Joris Vincent", email = "j.vincent@tu-berlin.de"},
 ]
-version = "0.101.0"
+version = "0.101.1"
 
 requires-python = ">=3.6"
 dependencies = [
