+++ .github/workflows/release.yml
@@ -13,7 +13,7 @@ jobs:
         uses: dsaltares/fetch-gh-release-asset@1.1.0
         with:
           regex: true
-          file: 'stimupy-.[0-9]*\.[0-9]*\.[0-9]*\.whl'
+          file: 'stimupy-.[0-9]*\.[0-9]*\.[0-9]*-py3-none-any\.whl'
           target: 'dist/'
 
       - name: Fetch sdist(s) from release
