### 2023-05-20 51b0d0e
build(release): publish to PyPI (& testPyPI), using trusted publisher
Having specified a "trusted publisher" (the repository and GHA workflow), no tokens and secrets are necessary.

BREAKING: Bump to v1.0.0
