### 2023-05-20 4c3c492
docs: update installation instructions to use PyPI
