### 2023-05-21 #5321 @LynnSchmittwilken
@JonathanReardon and @alexander-pastukhov, thank you for your detailed and your constructive feedback. The package and its documentation without a doubt improved as a result.

One final change to the manuscript: we added a paragraph to the manuscript in which we acknowledge the OCTA toolbox which is a Python toolbox to generate Gestalt-related visual stimuli
