### 2023-05-21 6bd55ae
fix (whites): if no index is provided, does not require target_heights anymore
