+++ stimupy/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.101.1"
+__version__ = "1.0.0"
 
 from stimupy import components, noises, stimuli, utils
 from stimupy.stimuli import *
