### 2023-05-27 #5321 @emdupre
Thank you again, @LynnSchmittwilken and @JorisVincent ! I've finished reviewing the code and accompanying paper, and I am also very happy with the current submission. There are a few editorial changes that I would request on the paper itself:

- Throughout, add a comma after "e.g."
- On L28, provide a citation for the first mention of ModelFest. I believe this would be the Carney and colleagues (1999) reference !
- On L32, add a space after the comma in "large,openly-available"
- On L49, "Matlab" should be in all capital letters
- On L63, change to " same—intuitive—parameters" to render as em-dashes
- On L70, "open science" should not be capitalized
- On L74, please use "NumPy" rather than "numpy" to standardize spelling throughout the document. Please also add a citation for the library
- On L102, please add a citation for Jupyter notebooks
- On L123, please move the examples outside of parentheses for readability. You could simply start this clause as "datasets; e.g., in spatial vision (Carney et al., 1999), ...."

And on the references:

- For `vincent.maertens.ea2021`, please update the reference to:
  ```
  @misc{vincent.maertens.ea2021,
  title={Vincent_LIG_Talk_2661: The missing linking functions in computational models of brightness perception},
  url={osf.io/9bca7},
  publisher={OSF},
  author={Vincent, Joris},
  year={2021},
  month={Sep}
  }
  ```
  to allow readers to direct readers to the associated OSF project.
- For `vincent.maertens2021`, please update to include this DOI, if correct: https://doi.org/10.1167/jov.22.3.20.
