### 2023-05-27 #5321 @emdupre
Following these changes, could you please:

- [ ] Check the archival deposit (e.g., in Zenodo) has the correct metadata (e.g., the JOSS paper and software archive have the same title and author listing)
- [ ] List the DOI of the archived version here.

I can then move forward with processing the submission 🚀
