+++ manuscript/paper.md
@@ -46,10 +46,10 @@ of different complexity as they are commonly used in the study of visual percept
 - different types of visual [noise](https://stimupy.readthedocs.io/en/latest/reference/_api/stimupy.noises.html) textures
 - visual [stimuli](https://stimupy.readthedocs.io/en/latest/reference/_api/stimupy.stimuli.html)
   such as Gabors, plaids, edges, and a variety of so-called illusions
-  (e.g. Simultaneous Brightness Contrast, White's illusion, Hermann grid, Ponzo illusion), and many more
-- stimulus sets from prior research papers, providing exact stimulus recreations (e.g. ModelFest)
+  (e.g., Simultaneous Brightness Contrast, White's illusion, Hermann grid, Ponzo illusion), and many more
+- stimulus sets from prior research papers, providing exact stimulus recreations (e.g., ModelFest, @carney1999)
 - [utility functions](https://stimupy.readthedocs.io/en/latest/reference/_api/stimupy.utils.html)
-  for stimulus import, export, manipulation (e.g. contrast, size), or plotting
+  for stimulus import, export, manipulation (e.g., contrast, size), or plotting
 - [documentation](https://stimupy.readthedocs.io/en/latest/index.html),
   including [interactive demonstrations](https://stimupy.readthedocs.io/en/latest/reference/demos.html) of stimulus functions
 - unit and integration [tests](https://github.com/computational-psychology/stimupy/actions/workflows/test.yml)
@@ -57,13 +57,13 @@ of different complexity as they are commonly used in the study of visual percept
 `stimupy` has been designed to:
 
 - generate (novel) visual stimuli in a reproducible, flexible, and easy way
-- recreate exact stimuli as they have been used in prior vision research 
+- recreate exact stimuli as they have been used in prior vision research
 - explore large parameter spaces to reveal relations between formerly unconnected stimuli
-- provide classic stimulus sets (e.g. ModelFest),
+- provide classic stimulus sets (e.g., ModelFest),
   exactly as described in the original manuscripts (including experimental data)
-- build new stimulus sets or benchmarks (e.g. for testing computational models),
+- build new stimulus sets or benchmarks (e.g., for testing computational models),
   and easily add them to `stimupy`
-- support vision science by providing a large,openly-available and flexible battery of relevant stimulus functions
+- support vision science by providing a large, openly-available and flexible battery of relevant stimulus functions
 - unify and automate stimulus creation
 
 ![A small fraction of the stimulus variety that ``stimupy`` can produce \label{fig:overview}](overview.png)
@@ -75,7 +75,7 @@ To generate stimuli, it is common practice to
 either write your own stimulus functions from scratch;
 reuse existing code;
 or import a static stimulus version from an image or data file
-(see e.g. @carney1999, @murray2020).
+(see e.g., @carney1999, @murray2020).
 The alternative to these idiosyncratic approaches
 is to use existing software which provides more flexible stimulus functions.
 
@@ -86,12 +86,12 @@ We are currently aware of
 - Pyllusion [@makowski2021],
 - OCTA [@OCTA].
 
-Psychtoolbox and Psychopy both provide functions to generate a number of visual stimuli. 
+Psychtoolbox and Psychopy both provide functions to generate a number of visual stimuli.
 However, stimulus generation is integrated into their main purpose
 which is to run psychophysical experiments.
 The design focus of both Psychtoolbox and Psychopy has therefore been
-to support the user to interface between computer hardware and Matlab and Python,
-respectively, 
+to support the user to interface between computer hardware and MATLAB and Python,
+respectively,
 to enable temporal precision and high dynamic range stimulus delivery.
 
 The design focus of `stimupy` is on stimulus creation.
@@ -100,7 +100,7 @@ than included in Psychtoolbox or Psychopy.
[…]
