### 2023-05-31 #5321 @emdupre
Yes, you can still change the manuscript title if necessary ! Please make sure that the software archive is listed with the same title.
