+++ manuscript/paper.md
@@ -1,5 +1,5 @@
 ---
-title: '`stimupy`: A Python stimulus creation package for vision science'
+title: '`stimupy`: A Python package for creating stimuli in vision science'
 tags:
   - Python package
   - vision science
