### 2023-06-05 2f37a9d
docs: final touch-up on manuscript
