+++ manuscript/bibliography.bib
@@ -84,7 +84,8 @@
 	editor = {F. Loizides and B. Schmidt},
 	organization = {IOS Press},
 	pages = {87 -- 90},
-	year = {2016}
+	year = {2016},
+	doi  = {10.3233/978-1-61499-649-1-87}
 } 
 
 @article{OCTA,
@@ -197,14 +198,16 @@
   copyright = {All rights reserved}
 }
 
-@misc{vincent.maertens.ea2021,
-  type    = {Talk},
+@inproceedings{vincent.maertens.ea2021,
   title   = {Perceptual Brightness Scales in a {{White}}'s Effect Stimulus Are Not Captured by Multiscale Spatial Filtering Models of Brightness Perception},
   author  = {Vincent, Joris and Maertens, Marianne and Aguilar, Guillermo},
-  year    = {2021},
-  month   = oct,
-  address = {{Online}},
-  langid  = {english}
+  year    = {2022},
+  month   = feb,
+  volume    = {22},
+  pages     = {20},
+  booktitle = {Journal of {{Vision}}},
+  langid  = {english},
+  doi     = {10.1167/jov.22.3.20}
 }
 
 @inproceedings{vincent.maertens.ea2022,
@@ -219,7 +222,7 @@
   doi       = {10.1167/jov.22.14.4160},
   urldate   = {2023-03-24},
   langid    = {english},
-  doi       = {10.1167/jov.22.3.20},
+  doi       = {10.1167/jov.22.14.4160},
 }
 
 @misc{vincent.maertens2021,
