### 2023-06-06 f57ec56
docs(manuscript): update references
