+++ manuscript/bibliography.bib
@@ -7,9 +7,7 @@
   volume    = {22},
   pages     = {3519},
   address   = {{St. Pete Beach, FL}},
-  doi       = {10.1167/jov.22.14.3519},
-  urldate   = {2023-03-24},
-  langid    = {english}
+  doi       = {10.1167/jov.22.14.3519}
 }
 
 @article{brainard1997,
@@ -56,36 +54,36 @@
   organization = {{IEEE}}
 }
 
-@Article{harris2020,
- title         = {Array programming with {NumPy}},
- author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
-                 van der Walt and Ralf Gommers and Pauli Virtanen and David
-                 Cournapeau and Eric Wieser and Julian Taylor and Sebastian
-                 Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
-                 and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
-                 Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
-                 R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
-                 G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
-                 Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
-                 Travis E. Oliphant},
- year          = {2020},
- journal       = {Nature},
- volume        = {585},
- number        = {7825},
- pages         = {357--362},
- doi           = {10.1038/s41586-020-2649-2},
- publisher     = {Springer Science and Business Media {LLC}},
+@article{harris2020,
+  title     = {Array programming with {NumPy}},
+  author    = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
+               van der Walt and Ralf Gommers and Pauli Virtanen and David
+               Cournapeau and Eric Wieser and Julian Taylor and Sebastian
+               Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
+               and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
+               Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
+               R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
+               G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
+               Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
+               Travis E. Oliphant},
+  year      = {2020},
+  journal   = {Nature},
+  volume    = {585},
+  number    = {7825},
+  pages     = {357--362},
+  doi       = {10.1038/s41586-020-2649-2},
+  publisher = {Springer Science and Business Media {LLC}}
 }
 
-@conference{kluyver2016,
-	title = {Jupyter Notebooks -- a publishing format for reproducible computational workflows},
-	author = {Thomas Kluyver and Benjamin Ragan-Kelley and Fernando P{\'e}rez and Brian Granger and Matthias Bussonnier and Jonathan Frederic and Kyle Kelley and Jessica Hamrick and Jason Grout and Sylvain Corlay and Paul Ivanov and Dami{\'a}n Avila and Safia Abdalla and Carol Willing},
-	booktitle = {Positioning and Power in Academic Publishing: Players, Agents and Agendas},
-	editor = {F. Loizides and B. Schmidt},
-	organization = {IOS Press},
-	pages = {87 -- 90},
-	year = {2016},
-	doi  = {10.3233/978-1-61499-649-1-87}
[…]
