### 2023-06-06 b46fdf8
build(manuscript): allow manual trigger
