### 2023-06-07 c3d999c
docs(manuscript): "use case" as two words
