### 2023-06-08 #5321 @JorisVincent
@emdupre Hm, this seems to be an issue on OSF's side. The OSF registration is marked as public, but not showing up as such (yet). We have contacted OSF about this, and hope to hear back from them soon.
