### 2023-06-09 #5321 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
